requirements.txt
langchain
openai
pinecone-client
langchain-pinecoin
langchain-openai
python-dotenv
pypdf
pytesseract